cbmir_ms_project/requirements.txt
streamlit>=1.32.0
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
pandas>=2.2.0
Pillow>=10.2.0
scikit-learn>=1.4.0
opencv-python-headless>=4.9.0
nibabel>=5.2.0
faiss-cpu>=1.8.0
matplotlib>=3.8.0
scipy>=1.12.0
tqdm>=4.66.0
